DbUtilの中での金額をintへ変更 年間集計・全帳簿・分類・項目集計の総収支に現金移動が含まれていたのを修正
git-svn-id: file:///Users/hiro/repository/beauties@73 b1584ab8-daee-4b4e-b8c6-5a3b018d5d45

+++ Beauties/src/model/Book.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Book {
-	private double mBalance = 0;
+	private int mBalance = 0;
 	private int mId;
 	private String mName;
 	
@@ -10,7 +10,7 @@ public class Book {
 		mName = pName;
 	}
 	
-	public void setBalance(double pBalance) {
+	public void setBalance(int pBalance) {
 		mBalance = pBalance;
 	}
 	
@@ -22,7 +22,7 @@ public class Book {
 		return mName;
 	}
 	
-	public double getBalance() {
+	public int getBalance() {
 		return mBalance;
 	}
